+++ lib/features/home_page/presentation/widgets/responsive_layout.dart
