+++ pubspec.yaml
@@ -9,6 +9,11 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  popover: ^0.3.0
+  flutter_bloc: ^8.1.4
+  equatable: ^2.0.5
+  get_it: ^7.6.7
+  dio: ^5.4.1
 
 dev_dependencies:
   flutter_test:
